[SIL] [Parser] Move SILParser into SILParser.h
Moves the declaration of SILParser into a new private header file: SILParser.h
